+++ opencog/CMakeLists.txt
@@ -22,6 +22,7 @@ ENDIF (HAVE_ATOMSPACE)
 
 IF (HAVE_SERVER)
     ADD_SUBDIRECTORY (dynamics)
+    ADD_SUBDIRECTORY (modules)
     ADD_SUBDIRECTORY (server)
 ENDIF(HAVE_SERVER)
