+++ opencog/modules/PersistModule.cc
