+++ opencog/query/CMakeLists.txt
@@ -32,13 +32,6 @@ IF (HAVE_GUILE)
 	)
 ENDIF (HAVE_GUILE)
 
-
-# Build the query cogserver module
-ADD_LIBRARY(QueryModule SHARED
-	QueryModule.cc
-)
-TARGET_LINK_LIBRARIES(QueryModule query server)
-
 IF (WIN32)
    INSTALL (TARGETS query DESTINATION "lib${LIB_DIR_SUFFIX}/opencog")
 ELSEIF (CYGWIN)
@@ -50,6 +43,7 @@ INSTALL (FILES
 	Instantiator.h
 	Pattern.h
 	PatternMatchCallback.h
+	PatternSCM.h
 	Satisfier.h
 	DESTINATION "include/${PROJECT_NAME}/query"
 )
